% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot state data on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number according to which state is selected}

\item{year}{The year to be selected.  This can be either string or numberic}
}
\value{
a graph with each selected observation ploted on the map
}
\description{
take state.num and year as input, extract longitude and latitude accordingly,
then plot the state on the map. Error will occur if state number is invalid.
}
\examples{
library(maps)
setwd(paste0(.libPaths()[1],"/Fas"))
fars_map_state(1, 2014)
}

