% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas_function.R
\name{fars_read}
\alias{fars_read}
\title{Read file as tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path of the file to be read}
}
\value{
tibble contain data in the file
}
\description{
Check whether file exists first, if it does, read the
the file then convert it to tibble format. Otherwise error message will return.
}
\examples{
setwd(paste0(.libPaths()[1],"/Fas"))
fars_read("accident_2013.csv.bz2")
}

