% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract data according to year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list contain years use to extract data, years can either denote as integers or strings.}
}
\value{
a list contain only the month and year provided, each element contain data for the same year. If the no data match the year provided NULL and a warning message will return.
}
\description{
The function take a list of years as input, use make_filename functon, and
fars_read to read data from disk according to the years provided. Only month
and year columns will be kept. Final output will be list format each element
contain data for one year.
}
\examples{
setwd(paste0(.libPaths()[1],"/Fas"))
years <- list(2014,2013)
fars_read_years(years)
}

