% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarize for years and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of years used for \code{fars_read_years} function.}
}
\value{
a table in tibble format, contain couts for each month-year combination.
}
\description{
summarize number count for each month and year combination
}
\examples{
years <- list("2014", 2015)
far_summrize_years(years)
}

