% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas_function.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename with provided year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a number or string represented the year that need to included in the filename.}
}
\value{
a string, represented the file name with pre-set format
}
\description{
Provide a year, this functon will generate the filename with pre-set format
 using the year provided. Thin input must be able to converted to integer or
 an error will occur.
}
\examples{
# all examples below generate same output
 make_filename(2014)
 make_filename("2014")
 make_filename(2014.3) #not recommended
}

